+++ server/src/services/shotspotterSensorsService.ts
@@ -1,0 +1,5 @@
+const { fetchShotspotterSensorsGeoJSON } = require('../repositories/shotspotterSensorsRepository');
+
+module.exports = {
+  getShotspotterSensorsGeoJSON: fetchShotspotterSensorsGeoJSON,
+};